requirements.txt
python-telegram-bot==20.3
Flask==2.0.1
apscheduler==3.9.1
python-dotenv==0.19.0